end proyect

+++ userPost/userPostCt.js
@@ -1,3 +1,4 @@
+const { verifyToken } = require("../usersData/config/handleJWT");
 const User = require("../usersData/userSchema/userSchema");
 const Post = require("./userPostMd");
 const controller = [];
@@ -14,36 +15,67 @@ controller.allPost = (req, res, next) => {
 };
 
 controller.createPost = async (req, res, next) => {
-  const { userName, fullName, content } = req.body;
-  const user = await User.find().where(userName);
-  let error = new Error("Usuario no encontrado");
+  if (req.headers.authorization) {
+    const { content } = req.body;
+    let token = req.headers.authorization.split(" ").pop();
+    let validate = await verifyToken(token);
 
-  console.log("este es el log del user => ", user[0]);
-  if (!user.length) {
-    error.status = 401;
-    return next(error);
-  }
-
-  const newPost = new Post({
-    userName,
-    fullName,
-    content,
-  });
-
-  await newPost.save((error) => {
-    if (error) {
-      error.status = 500;
-      next(error);
-    } else {
-      res.status(200).json(newPost);
+    if (validate instanceof Error) {
+      return res.status(403).json({ message: validate });
     }
-  });
+
+    const user = await User.find().where(validate.username);
+    let error = new Error("Usuario no encontrado");
+
+    if (!user.length) {
+      error.status = 401;
+      return next(error);
+    }
+
+    const newPost = new Post({
+      userName: validate.username,
+      fullName: validate.fullName,
+      content,
+    });
+
+    await newPost.save((error) => {
+      if (error) {
+        error.status = 500;
+        next(error);
+      } else {
+        res.status(200).json(newPost);
+      }
+    });
+  } else {
+    res.status(500).json({ message: "Invalid or not found token" });
+  }
 };
 
 controller.deletePost = async (req, res, next) => {
   try {
-    const deletedUser = await Post.findByIdAndDelete(req.params.id);
-    res.status(200).json({ message: deletedUser, status: 200 });
+    if (req.headers.authorization) {
+      let token = req.headers.authorization.split(" ").pop();
+      let validate = await verifyToken(token);
+
+      if (validate instanceof Error) {
+        return res.status(403).json({ message: validate });
+      }
+
+      const PostUser = await Post.findById(req.params.id);
+      if (!PostUser) {
+        return res.status(404).json({ message: "Post not found " });
+      }
+
+      if (PostUser.userName == validate.username) {
+        const deletedPost = await Post.findByIdAndDelete(req.params.id);
+        return res.status(200).json({ message: deletedPost, status: 200 });
+      }
+      return res
+        .status(500)
+        .json({ message: "This post does not belong to the user" });
+    } else {
+      res.status(500).json({ message: "Invalid or not found token" });
+    }
   } catch (error) {
     error.status = 500;
     next(error);
